requirements.txt
pandas
numpy
pyarrow
hdbscan
scikit-learn
python-telegram-bot
python-dotenv
openai
matplotlib
aiogram